Modified Translater to format iam roles for backward compatibility
This CR contains changes in schema.json for namespace resource. This
schema change adds the necessary policy changes to support resource
policy and aws secrets manager integration for namespace resource. This
CR also contains change for IAMRoles translator which is currently
being returned differently by RACS. We have used regex for backward
compatability as the CT expect it to be list of strings with just IAM
role ARN.

Testing:
1. Run changes to test handler locally
    a. source <pythonenv>/bin/activate && aws configure
    b. cd <resource> && sam local start-lambda
    c. cfn invoke resource CREATE <payload.json>

2. Run CT in personal account
    a. aws s3 cp build/packaging_additional_published_artifacts/aws-redshiftserverless-namespace.zip s3://vramanet-cfn-test-local
    b. docker run --env AWS_ACCESS_KEY_ID=$AWS_ACCESS_KEY_ID \ --env AWS_SECRET_ACCESS_KEY=$AWS_SECRET_ACCESS_KEY \ --env AWS_SESSION_TOKEN=$AWS_SESSION_TOKEN \ --env AWS_DEFAULT_REGION=us-west-2 \ -p 9000:8080 public.ecr.aws/j9c3o4f9/contract-tests:latest
    c. curl -XPOST "http://localhost:9000/2015-03-31/functions/function/invocations" -d '{ "TypeName": "AWS::RedshiftServerless::Namespace", "Bucket": "vramanet-cfn-test-local", "Key": "aws-redshiftserverless-namespace.zip" }'