Catch all exceptions when resourcePolicy is not present
Our CFN handler needs getResourcePolicy permission to see if
resource-policy is created or not, but at the same time we
need to support existing customers who don't have this permission,
so we added some error catching logic to be backward compatible.

Turns out the previous exception catching logic is not sufficient to
catch all customer exceptions.

This commit catches all exceptions when resourcePolicy is not present,
to unblock all existing customers who are not using resourcePolicy
issue. Our team will evaluate and restrict the catching logic if needed