Retry create/update on workgroup conflict exception
In eu-north-1 region, some wf steps take longer to finish, and contract
test would fail because of ConflictException("There is an operation
running on the workgroup") and contract test will fail.

This commit handles the exception by matching the exception message and
retry on the exception, giving the wf more time to finish.