Fix for Workgroup stabilization failures on operations
Description: This CR contains fix for stabilization not correctly
implemented for Workgroup Resource. During Read handler we are returing
in Progress status which is not allowed. This CR will fix that issue.

Currently as part of create and delete handlers we are only checking if
Workgroup is stablized. This is causing failures if the same workgroup
is created immediately causing "Operation in progress failure". This is
because Namespace is not stabilized during the operation. This CR fixes
that issue as well.

Testing:
1. Tested if CTV2 is passing in eu-north-1 and us-west-1 region which
   are currently failing CTV2 for Workgroup resource.
2. Verified CTV2 is passing for those regions in personal account.