Description: This CR contains fix for stabilization in delete handler for Workgroup Resource. As part of CTV2 we will delete a resource and create the resource immediately in next run. This will cause CTV2 to fail for workgroup.
Currently namespace stabilization is checked only during create and not during delete. This is causing failures if the same workgroup is created after delete from earlier run immediately causing "Operation in progress failure". This is because Namespace is not stabilized during the operation. This CR fixes that issue during delete stabilization

Testing:

    Tested if CTV2 is passing in ap-northeast-2 and us-west-1 region which are currently failing CTV2 for Workgroup resource.
    Verified CTV2 is passing for those regions in personal account.